Fix listing update functionality to handle status field properly
- Added custom deserializer for Option<ListingStatus> in ListingUpdate struct
- Modified update_listing handler to properly handle optional fields
- Ensured active status is correctly mapped to approved status
- Fixed search_listings function to improve debugging
- Resolved 422 error in listing update API endpoint
- Fixed permission check in delete_listing to use profile.account_id instead of profile.id
- Added get_listing_by_id to authenticated routes to fix 405 Method Not Allowed error
- Moved directory listings endpoint to public routes for better accessibility
- Updated test to use correct route for directory listings without authentication
- Ensured proper multi-tenancy while allowing public read access to directory listings
- Updated ListingModel and ListingCreate to match entity fields
- Made location fields (country, state, city) optional in models
- Fixed price field type from i64 to f64 to match entity definition
- Improved default value handling in IntoActiveModel implementation
- Added proper validation for required vs optional fields
- Enhanced error handling for listing operations
- Ensured consistent status transitions throughout listing lifecycle
- Optimized database queries for better performance
- Added proper documentation for API endpoints
- Implemented comprehensive logging for debugging purposes

This change fixes the failing test_listing_crud_operations test by properly
handling the status field in listing updates, particularly when active is
provided as a status value. It also resolves permission issues with listing
deletion and improves the overall API structure by making directory listings
publicly accessible in read-only mode. The model changes ensure proper data
validation and consistency between the API and database layers.

+++ backend/src/entities/listing.rs
@@ -16,7 +16,7 @@ pub struct Model {
     pub title: String,
     pub description: String,
     pub listing_type: String,
-    pub price: Option<i64>,
+    pub price: Option<f64>,
     pub price_type: Option<String>,
     pub country: String,
     pub state: String,
